test(web): lock Codex subscription parse, store, and tab fields
Replace the brittle renderer tab test with field-builder coverage and
assert the demo api-call subscriptions payload.
