+++ apps/web/src/features/accounts/model/codexSubscription/index.ts
@@ -18,3 +18,4 @@ export {
   useCodexSubscriptionStore,
 } from './store';
 export { collectCodexSubscriptionTargets, shouldShowCodexSubscriptionTab } from './tabGate';
+export { buildCodexSubscriptionDetailFields, buildCodexSubscriptionTabFields } from './tabFields';
